settime: Replace int with ssize_t
Because the write pointer type is different from the function define,
there is a warning information caused by a35378a as follows:
notify_settime.c:99:2: warning: initialization from incompatible pointer type
  .write = notify_settime_signal_proc_write,
    ^
For fixing it, Replace int with ssize_t.

+++ kernel/notify_settime.c
@@ -73,7 +73,7 @@ static int notify_settime_signal_proc_open(struct inode *inode, struct file *fil
 	return single_open(file, notify_settime_signal_show, PDE_DATA(inode));
 }
 
-static int notify_settime_signal_proc_write(struct file *file,
+static ssize_t notify_settime_signal_proc_write(struct file *file,
 					const char __user *buffer,
 				       size_t count, loff_t *pos)
 {
